backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==42.0.2
cssselect2==0.7.0
Django==5.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-environ==0.11.2
django-idmap==1.0.3
django-qr-code==4.0.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
html5lib==1.1
idna==3.6
kombu==5.3.7
lxml==4.9.4
numpy==1.26.3
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.3.0.post0
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.0.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==3.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
redis==5.0.4
reportlab==4.0.9
requests==2.31.0
segno==1.6.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
trio==0.23.2
trio-websocket==0.11.1
trycourier==5.0.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wsproto==1.2.0
xhtml2pdf==0.2.15
XlsxWriter==3.2.0
